+++ tests/arg/test_invalid_annotation_combination.py
@@ -31,7 +31,7 @@ def test_sequence_unioned_with_scalar(backend):
 def test_sequence_with_scalar_action(backend):
     @dataclass
     class Args:
-        foo: Annotated[list[str], cappa.Arg(action=cappa.ArgAction.set)]
+        foo: Annotated[list[str], cappa.Arg(action=cappa.ArgAction.set, num_args=1)]
 
     with pytest.raises(ValueError) as e:
         parse(Args, "--help", backend=backend)
@@ -50,13 +50,22 @@ def test_sequence_with_scalar_num_args(backend):
     class Args:
         foo: Annotated[list[str], cappa.Arg(num_args=1, short=True)]
 
+    args = parse(Args, "-f", "a", "-f", "b", backend=backend)
+    assert args == Args(["a", "b"])
+
+    @dataclass
+    class ArgsBad:
+        foo: Annotated[
+            list[str], cappa.Arg(num_args=1, short=True, action=cappa.ArgAction.set)
+        ]
+
     with pytest.raises(ValueError) as e:
-        parse(Args, "--help", backend=backend)
+        parse(ArgsBad, "--help", backend=backend)
 
     result = str(e.value).replace("typing.List", "list")
     assert result == (
         "On field 'foo', apparent mismatch of annotated type with `Arg` options. "
-        "'list[str]' type produces a sequence, whereas `num_args=1`/`action=None` do not. "
+        "'list[str]' type produces a sequence, whereas `num_args=1`/`action=ArgAction.set` do not. "
         "See [documentation](https://cappa.readthedocs.io/en/latest/annotation.html) for more details."
     )